requirements.txt
# ── Web framework ─────────────────────────────────────────────────────────────
fastapi==0.129.0
uvicorn==0.40.0
gunicorn==21.2.0
starlette==0.52.1
python-multipart==0.0.22

# ── Database ──────────────────────────────────────────────────────────────────
SQLAlchemy==2.0.46
alembic==1.18.4
psycopg2-binary==2.9.11
greenlet==3.3.1

# ── Pydantic / validation ─────────────────────────────────────────────────────
pydantic==2.12.5
pydantic_core==2.41.5
annotated-types==0.7.0
typing_extensions==4.15.0
typing-inspection==0.4.2

# ── Auth / JWT ────────────────────────────────────────────────────────────────
python-jose==3.5.0
passlib==1.7.4
cryptography==46.0.5
ecdsa==0.19.1
pyasn1==0.6.2
rsa==4.9.1
PyJWT==2.10.1

# ── Environment / config ──────────────────────────────────────────────────────
python-dotenv==1.2.1
python-decouple==3.8

# ── HTTP / networking ─────────────────────────────────────────────────────────
requests==2.32.5
urllib3==2.6.3
certifi==2026.1.4
charset-normalizer==3.4.4
idna==3.11
h11==0.16.0
anyio==4.12.1

# ── Excel (import/export catálogo) ─────────────────────────────────────────────
pandas==2.2.3
openpyxl==3.1.5

# ── Utils ─────────────────────────────────────────────────────────────────────
click==8.3.1
colorama==0.4.6
packaging==25.0
six==1.17.0
Mako==1.3.10
MarkupSafe==3.0.3
tzdata==2025.2
cffi==2.0.0
pycparser==3.0
